Aahaar - Update
In this update, a lot of new improvements were added. The Vendor page in the admin now is able to correctly register a new vendor, and create his sub account on Razorpay to enable him for route transsactions. Still needs debugging, and the view page and edit page need to be made too.