A lot of improvement to Aahaar
Nearing the Testing, All the dashboards are being updated to use Razorpay for the payement. Animations are being improved to feel user centric.
UI is being updated to reflect the motifs of the Brand.
The data flow is also being updated to be consistent and secure.
Application is nearing its public release.

Hoping for the best!